backend/requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Pydantic & Settings
pydantic==2.7.1
pydantic-settings==2.3.0

# Database (async)
sqlalchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0
# For PostgreSQL in production, add:
# asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ML & Data Science
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
lightgbm==4.3.0
shap==0.45.1
joblib==1.4.2

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
